feat: use int instead of size_t

+++ sort/bubble_sort.cpp
@@ -1,13 +1,11 @@
-#include <cstddef>
-#include <utility>
-#include <vector>
+#include <bits/stdc++.h>
 
 using namespace std;
 
 void bubble_sort(vector<int> &a)
 {
-	for (size_t i = a.size() - 1; i >= 1; i--) {
-		for (size_t j = 0; j < i; j++) {
+	for (int i = a.size() - 1; i >= 1; i--) {
+		for (int j = 0; j < i; j++) {
 			if (a[j] > a[j + 1]) {
 				swap(a[j], a[j + 1]);
 			}
