+++ sort/selection_sort.cpp
@@ -1,9 +1,8 @@
-#include <cstddef>
-#include <vector>
+#include <bits/stdc++.h>
 
 using namespace std;
 
-size_t get_index_of_first_min(const vector<int> &a, size_t left, size_t right)
+int get_index_of_first_min(const vector<int> &a, int left, int right)
 {
 	if (left > right) {
 		return 0;
@@ -12,7 +11,7 @@ size_t get_index_of_first_min(const vector<int> &a, size_t left, size_t right)
 	int min = a[left];
 	int min_index = left;
 
-	for (size_t i = left + 1; i <= right; i++) {
+	for (int i = left + 1; i <= right; i++) {
 		if (a[i] < min) {
 			min = a[i];
 			min_index = i;
@@ -24,8 +23,8 @@ size_t get_index_of_first_min(const vector<int> &a, size_t left, size_t right)
 
 void selection_sort(vector<int> &a)
 {
-	for (size_t i = 0; i < a.size(); i++) {
-        size_t min_index = get_index_of_first_min(a, i , a.size() - 1);
+	for (int i = 0; i < a.size(); i++) {
+        int min_index = get_index_of_first_min(a, i , a.size() - 1);
         swap(a[i], a[min_index]);
 	}
 }
